Changed GraphQLControllerTests to GraphQLControllerMockDataTests and put all mock data tests in this file. Renamed GraphQLMutationTests to GraphQLControllerTests in preparation to put tests with live network calls.
